created login view
