+++ accounts/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+from django.contrib.auth.views import LoginView, LogoutView
+
+urlpatterns = [
+    path("login", LoginView.as_view(), name="login"),
+    path("logout", LogoutView.as_view(), name="logout")
+]
