+++ archives/settings.py
@@ -38,7 +38,8 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
 
-    'api.apps.ApiConfig'
+    'api.apps.ApiConfig',
+    'accounts.apps.AccountsConfig'
 ]
 
 MIDDLEWARE = [
